feat: initial
Having issues passing openapi validation...but the swagger UI is visible.

**notes**

> Ability to find all Dielines that have a certain number of shapes (integer)

Use a deep object for this

> Ability to find all Dielines that match a given shape (path object)

Assuming the svg(?) path object could potentially be hugely complex, it's too big for a query string. So, assume that the database also persists the shapes and we can reference them via id.

> Ability to restore a given version of a Dieline object as the latest (current) version

Should be doable just by adding a `is_current` flag on the entity, the `PUT /dieline/{dieline_id}` endpoint can set this flag.